+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Fixed long tooltips text by restricting them to 100 characters. (#189)
   The text in the various viewing dialogs is still the full length.
   The 100 limit probably needs to be configurable.
+* Added register profiles to show only interesting/relevant registers.
+* Added UTF-8,16,32 support in the Memory Visualizer.
 
 ## [2.3] - 2023-11-19
 * In the margins of the source windows, allow CTRL+DoubleClick to do a quick RunToLine or RunToAddress.